requirements.txt
appdirs==1.4.4
arrow==1.3.0
binaryornot==0.4.4
Brotli==1.1.0
build==1.2.1
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy
cookiecutter==2.6.0
cycler==0.12.1
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
einops==0.7.0
evalutils==0.4.2
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
gmpy2==2.1.2
huggingface-hub==0.20.3
idna==3.6
imagecodecs
imageio==2.33.1
imageio-ffmpeg==0.4.9
imgaug==0.4.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning==2.1.3
lightning-utilities==0.10.1
line_profiler==4.1.3
llvmlite
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib
mdurl==0.1.2
mpmath==1.3.0
munch==4.0.0
munkres==1.1.4
numba
numpy
opencv-python
packaging==23.2
pandas
pathtools==0.1.2
pillow==10.2.0
pip==23.3.2
pip-tools==7.4.1
pretrainedmodels==0.7.4
protobuf==4.25.1
psutil==5.9.8
Pygments==2.18.0
pyparsing==3.1.1
pyproject_hooks==1.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.4
pytorch-lightning==2.1.3
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
safetensors==0.4.2
scikit-image
scikit-learn
scipy
segmentation-models-pytorch
sentry-sdk==1.40.2
setproctitle==1.3.3
setuptools==69.0.3
shapely==2.0.2
SimpleITK==2.4.0
six==1.16.0
smmap==5.0.0
sympy==1.12
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile
timm
# torch==2.1.2
# torchaudio==2.1.2
# torchinfo==1.8.0
# torchmetrics==1.2.1
# torchsummary==1.5.1
# torchvision==0.16.2
tornado==6.3.3
tqdm==4.66.1
types-python-dateutil==2.9.0.20240821
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
wandb==0.16.3
wheel==0.42.0
yacs==0.1.8
